New file MainState
